\name{blkdiag}
\alias{blkdiag}
\title{diagonal stack of block matrices}
\description{
 Build a block diagonal matrix
}
\usage{
blkdiag(...)
}
\arguments{
  \item{...}{sequence of non-empty, numeric matrices}
}
\details{
Build a block diagonal matrix. Building block matrices do not
need to be square, but need to be numeric and non-empty matrices
}
\value{
  A matrix
}
\author{Javier Garcia-Pintado, e-mail:
  \email{j.garcia-pintado@reading.ac.uk}}
\keyword{ utilities }
\keyword{ spatial }
