\name{hypot}
\alias{hypot}
\title{Calculate hypotenuse}
\description{Calculate hypotenuse given the lengths of the other two sides
}
\usage{
 hypot(a, b)
}
\arguments{
  \item{a}{a side length}
  \item{b}{the other side length}
}
\value{
  A scalar or vector of hypotenuse length/s
}
\author{Javier Garcia-Pintado, e-mail: \email{j.garcia-pintado@reading.ac.uk}}
\keyword{ utilities }
\keyword{ spatial }
