\name{linearScale}
\alias{linearScale}
\title{ Linear Scale Transformation}
\description{
 Conduct a simple linear scale as transformation example
}
\usage{
linearScale(x, a=0.0, b=1.0, inverse=FALSE)
}
\arguments{
  \item{x}{REAL, vector to be transformed}
  \item{a}{additive factor in linear scale transformation}
  \item{b}{multiplicative factor in linear scale transformation}
  \item{inverse}{LOGICAL, wether the transformation is inverse}
}
\details{
 Linear scaling as example of parametric transformation function
}
\value{
  A vector
}
\keyword{ utilities }
