\name{qpow35}
\alias{qpow35}
\title{ Powerlaw Transformation}
\description{
 Conduct a powerlaw as transformation 
}
\usage{
qpow35(x, inverse=FALSE)
}
\arguments{
  \item{x}{REAL, vector to be transformed}
  \item{inverse}{LOGICAL, wether the transformation is inverse}
}
\details{
 Powerlaw scaling as example of parametric transformation function
}
\value{
  A vector
}
\keyword{ utilities }

