\name{rDAF-package}
\alias{rDAF-package}
\alias{rDAF}
\docType{package}
\title{
\packageTitle{rDAF}
}
\description{
\packageDescription{rDAF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rDAF}
\packageIndices{rDAF}
A software for data assimilation based on R.
The current version only provides the core function assimilateLite().
A version including localization, inflation, rotation, and MPI
parallelization is under testing and being documented. Please contact
the author for information requests.
}
\author{
\packageAuthor{rDAF}

Maintainer: \packageMaintainer{rDAF}
}
\keyword{ package }

