\name{rDAF-package}
\alias{rDAF-package}
\alias{rDAF}
\docType{package}
\title{
\packageTitle{rDAF}
}
\description{
\packageDescription{rDAF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rDAF}
\packageIndices{rDAF}
A software for data assimilation based on R. Localization routines are
under development. In the current (lite) version, the organization of the
matrices needed to conduct the assimilation is given, but only
assimilateLite() is provided as workhorse assimilation function, which
does not include any localization, inflation, rotation, or MPI parallelization.
}
\author{
\packageAuthor{rDAF}

Maintainer: \packageMaintainer{rDAF}
}
\keyword{ package }

